Log filename in frontend error logging
changelog: Internal, Error Logging, Log filename in frontend error logging
