+++ app/javascript/packages/analytics/track-event.spec.ts
@@ -1,4 +1,4 @@
-import { trackEvent, trackError } from '@18f/identity-analytics';
+import { trackEvent } from '@18f/identity-analytics';
 import { useSandbox } from '@18f/identity-test-helpers';
 import type { SinonStub } from 'sinon';
 
@@ -80,32 +80,5 @@ describe('trackEvent', () => {
         expect(result).to.be.undefined();
       });
     });
-  });
-});
-
-describe('trackError', () => {
-  const sandbox = useSandbox();
-  const endpoint = '/log';
-
-  beforeEach(() => {
-    sandbox.stub(global.navigator, 'sendBeacon').returns(true);
-    document.body.innerHTML = `<script type="application/json" data-config>{"analyticsEndpoint":"${endpoint}"}</script>`;
-  });
-
-  it('tracks event', async () => {
-    trackError(new Error('Oops!'));
-
-    expect(global.navigator.sendBeacon).to.have.been.calledOnce();
-
-    const [actualEndpoint, data] = (global.navigator.sendBeacon as SinonStub).firstCall.args;
-    expect(actualEndpoint).to.eql(endpoint);
-
-    const { event, payload } = JSON.parse(await data.text());
-    const { name, message, stack } = payload;
-
-    expect(event).to.equal('Frontend Error');
-    expect(name).to.equal('Error');
-    expect(message).to.equal('Oops!');
-    expect(stack).to.be.a('string');
   });
 });
