+++ app/javascript/packs/track-errors.ts
@@ -9,6 +9,6 @@ declare let window: WindowWithInitialErrors;
 const { _e: initialErrors } = window;
 
 const handleErrorEvent = (event: ErrorEvent) =>
-  isTrackableErrorEvent(event) && trackError(event.error);
+  isTrackableErrorEvent(event) && trackError(event.error, event);
 initialErrors.forEach(handleErrorEvent);
 window.addEventListener('error', handleErrorEvent);
